refactor: :recycle: remove unnecessary container from TasksList
