+++ src/components/tasks/TasksList/TasksList.tsx
@@ -15,10 +15,8 @@ import isEmpty from 'lodash/isEmpty';
 import React, { useState } from 'react';
 import { When } from 'react-if';
 import { Link } from 'react-router-dom';
-import { useTypedSelector } from '../../../store/index';
-import { Task } from "../../../entities/Task";
-import { tasksSelector } from '../../../store/selectors';
 import { tasksPerPage } from '../../../contants';
+import { Task } from '../../../entities/Task';
 
 const log = debug('TasksList');
 interface TasksListProps {
@@ -138,17 +136,4 @@ function useStyles() {
       },
     };
   })();
-}
-
-// TODO add props types
-export default function TasksListContainer(props: any) {
-  const tasks = useTypedSelector(tasksSelector);
-
-  const mergeProps = {
-    ...props,
-    loading: tasks === undefined,
-    tasks,
-  };
-
-  return <TasksList {...mergeProps} />;
 }
