+++ src/components/tasks/TasksList/index.ts
@@ -1,0 +1,1 @@
+export * from './TasksList';
